Modified finance_calculator.py

+++ python_introduction/finance_calculator.py
@@ -3,8 +3,8 @@ name = input("Hey what's your name? ")
 print("Hello!",name,)
 print("Welcome to the finance management program.\nMay I ask what you would like to start with?")
 
-monthly_income = int(input("Enter your monthly income: "))
-monthly_expense = int(input("Enter your total monthly expense: "))
+monthly_income = float(input("Enter your monthly income: "))
+monthly_expense = float(input("Enter your total monthly expense: "))
 monthly_savings =  float(monthly_income) - float(monthly_expense)
 
 print("Your monthly savings are kes",monthly_savings)
